fix: fix ProtectiveRoute & refactor codebase
